Adding DataModels and renaming namespace to ROSNET
