+++ ROSNET/DataModel/IRecord.cs
@@ -1,0 +1,8 @@
+﻿using System;
+namespace ROSNET.DataModel
+{
+    public interface IRecord
+    {
+       
+    }
+}
