+++ ROSNET/DataModel/ROSbag.cs
@@ -1,0 +1,11 @@
+﻿using System;
+namespace ROSNET.DataModel
+{
+    public class ROSbag
+    {
+        public ROSbag()
+        {
+
+        }
+    }
+}
