+++ ROSNET/ROSNET.csproj
@@ -5,4 +5,10 @@
     <RootNamespace>ROSbag_ReadWrite</RootNamespace>
   </PropertyGroup>
 
+  <ItemGroup>
+    <None Remove="DataModel\" />
+  </ItemGroup>
+  <ItemGroup>
+    <Folder Include="DataModel\" />
+  </ItemGroup>
 </Project>
